assignments/group_projects/youtube_summarizer/requirements.txt
google-cloud-speech
youtube-dl
